requirements.txt
alabaster==0.7.8
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
astroid==2.3.3
asttokens==2.0.5
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
autopep8==1.6.0
Babel==2.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
black==22.3.0
bleach==5.0.0
blinker==1.4
bokeh==2.4.3
Brlapi==0.7.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==1.3.0
cmake==3.22.4
colorama==0.4.3
command-not-found==0.3
configparser==5.2.0
constantly==15.1.0
coverage==6.4
cryptography==36.0.2
cupshelpers==1.0
cycler==0.11.0
daphne==3.0.2
dbus-python==1.2.16
debugpy==1.6.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.3
django-allauth==0.50.0
django-countries==7.3.2
django-decouple==2.1
django-environ==0.8.1
django-js-asset==2.0.0
django-markdown==0.8.4
django-model-utils==4.2.0
django-mptt==0.13.4
django-notifications-hq==1.7.0
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
django-rest-framework==0.1.0
django-summernote==0.8.20.0
django-taggit==2.1.0
django-tinymce==3.4.0
djangorestframework==3.13.1
dlib==19.24.0
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
executing==0.8.3
face-recognition==1.3.0
face-recognition-models==0.3.0
fasteners==0.14.1
fastjsonschema==2.15.3
filelock==3.0.12
flake8==4.0.1
Flask==2.1.2
fonttools==4.32.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.0.1
httplib2==0.14.0
hyperlink==21.0.0
idna==3.3
imagesize==1.2.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
incremental==21.3.0
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==4.3.4
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
jsonfield==3.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keyring==18.0.1
kiwisolver==1.4.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
logilab-common==1.4.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==2.1.1
martor==1.6.13
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
monotonic==1.5
more-itertools==4.2.0
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysql-connector-python==8.0.29
mysqlclient==2.1.0
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
netifaces==0.10.4
notebook==6.4.10
numpy==1.22.3
numpydoc==0.7.0
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pathspec==0.9.0
pbr==5.4.5
pexpect==4.6.0
phonenumbers==8.12.48
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.6.1
psutil==5.9.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.21
pycups==1.9.73
pyflakes==2.4.0
pygame==2.1.2
Pygments==2.11.2
PyGObject==3.36.0
PyJWT==2.3.0
pylint==2.4.4
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.18.1
pysvn==1.9.9
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
QtAwesome==0.4.4
qtconsole==5.3.0
QtPy==2.0.1
razorpay==1.3.0
reportlab==3.5.34
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
roman==2.0.0
rope==0.10.5
scikit-learn==1.0.2
scipy==1.8.0
SecretStorage==2.3.1
Send2Trash==1.8.0
service-identity==21.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sklearn==0.0
soupsieve==2.3.2.post1
Sphinx==1.8.5
spyder==3.3.6
spyder-kernels==0.5.2
sqlparse==0.4.2
stack-data==0.2.0
swapper==1.3.0
systemd-python==234
terminado==0.13.3
testpath==0.4.4
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.9
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.0
wrapt==1.11.2
xkit==0.0.0
yapf==0.32.0
zipp==3.8.0
zope.interface==5.4.0
